Add confirmation prompt to DeleteCommands ('are you sure you want to delete this?').

+++ cloudCacheCLI/Commands/DeleteCommand.py
@@ -3,6 +3,7 @@
 import requests
 
 from . import BaseCommand
+from distutils.util import strtobool
 
 # -------------------------------------------------------------------------------------------------
 
@@ -17,10 +18,15 @@ class DeleteCommand(BaseCommand):
     def action(self):
         """ Evaluates this Command by performing its API call. The response object itself, and the json/dict contents
         of the response, are set as instance attributes so we can reference them later. """
-        self.response = requests.delete(self.url, headers=self.headers)
-        super(DeleteCommand, self).action()
+
+        prompt = '\n{}\nEnter `yes` or `no` (or `y` or `n`): '.format(self.prompt)
+        user_confirmation = bool(strtobool(input(prompt)))
+
+        if user_confirmation:
+            self.response = requests.delete(self.url, headers=self.headers)
+            super(DeleteCommand, self).action()
 
 
     def _on_action_success(self):
         """ OVERRIDE - Do nothing, the delete action was successful. """
-        pass
+        print('\nSuccessfully deleted.')
